requirements.txt
python-dotenv==1.0.1
psycopg[binary]==3.2.6
google-ads-admanager==0.5.0
streamlit==1.48.1
pandas==2.3.2
plotly==6.3.0
